shuffling around specs
